#3070 - Modularize webapp module
- Move IO supports for LIF, NIF, IMSCWB and Perseus into separate modules
