+++ inception/inception-io-lif/src/main/java/de/tudarmstadt/ukp/inception/io/lif/LifFormatSupport.java
@@ -15,24 +15,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.formats;
+package de.tudarmstadt.ukp.inception.io.lif;
 
+import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
 
+import org.apache.uima.analysis_engine.AnalysisEngineDescription;
+import org.apache.uima.cas.CAS;
 import org.apache.uima.collection.CollectionReaderDescription;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.dkpro.core.io.imscwb.ImsCwbReader;
+import org.dkpro.core.io.lif.LifReader;
+import org.dkpro.core.io.lif.LifWriter;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
+import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 
 @Component
-public class ImsCwbFormatSupport
+public class LifFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "imscwb";
-    public static final String NAME = "Corpus Workbench Format (aka VRT)";
+    public static final String ID = "lif";
+    public static final String NAME = "LAPPS Interchange Format";
 
     @Override
     public String getId()
@@ -52,23 +57,24 @@ public class ImsCwbFormatSupport
         return true;
     }
 
-    // @Override
-    // public boolean isWritable()
-    // {
-    // return true;
-    // }
+    @Override
+    public boolean isWritable()
+    {
+        return true;
+    }
 
     @Override
     public CollectionReaderDescription getReaderDescription(TypeSystemDescription aTSD)
         throws ResourceInitializationException
     {
-        return createReaderDescription(ImsCwbReader.class, aTSD);
+        return createReaderDescription(LifReader.class, aTSD);
     }
 
-    // @Override
-    // public AnalysisEngineDescription getWriterDescription(Project aProject, CAS aCAS)
-    // throws ResourceInitializationException
-    // {
-    // return createEngineDescription(ImsCwbWriter.class);
-    // }
+    @Override
+    public AnalysisEngineDescription getWriterDescription(Project aProject,
+            TypeSystemDescription aTSD, CAS aCAS)
+        throws ResourceInitializationException
+    {
+        return createEngineDescription(LifWriter.class, aTSD);
+    }
 }
