+++ inception/inception-io-perseus/src/main/java/de/tudarmstadt/ukp/inception/io/perseus/PerseusFormatSupport.java
@@ -15,29 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.formats;
+package de.tudarmstadt.ukp.inception.io.perseus;
 
-import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
 
-import org.apache.uima.analysis_engine.AnalysisEngineDescription;
-import org.apache.uima.cas.CAS;
 import org.apache.uima.collection.CollectionReaderDescription;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.dkpro.core.io.lif.LifReader;
-import org.dkpro.core.io.lif.LifWriter;
+import org.dkpro.core.io.perseus.PerseusReader;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
-import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 
 @Component
-public class LifFormatSupport
+public class PerseusFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "lif";
-    public static final String NAME = "LAPPS Interchange Format";
+    public static final String ID = "perseus_2.1";
+    public static final String NAME = "Perseus treebank XML (2.1)";
 
     @Override
     public String getId()
@@ -60,21 +55,13 @@ public class LifFormatSupport
     @Override
     public boolean isWritable()
     {
-        return true;
+        return false;
     }
 
     @Override
     public CollectionReaderDescription getReaderDescription(TypeSystemDescription aTSD)
         throws ResourceInitializationException
     {
-        return createReaderDescription(LifReader.class, aTSD);
-    }
-
-    @Override
-    public AnalysisEngineDescription getWriterDescription(Project aProject,
-            TypeSystemDescription aTSD, CAS aCAS)
-        throws ResourceInitializationException
-    {
-        return createEngineDescription(LifWriter.class, aTSD);
+        return createReaderDescription(PerseusReader.class, aTSD);
     }
 }
