+++ inception/inception-io-xmi/src/main/java/de/tudarmstadt/ukp/clarin/webanno/xmi/UimaInlineXmlFormatSupport.java
@@ -15,24 +15,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.formats;
+package de.tudarmstadt.ukp.clarin.webanno.xmi;
 
-import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
+import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 
-import org.apache.uima.collection.CollectionReaderDescription;
+import org.apache.uima.analysis_engine.AnalysisEngineDescription;
+import org.apache.uima.cas.CAS;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.dkpro.core.io.perseus.PerseusReader;
+import org.dkpro.core.io.xml.InlineXmlWriter;
 import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
+import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 
 @Component
-public class PerseusFormatSupport
+public class UimaInlineXmlFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "perseus_2.1";
-    public static final String NAME = "Perseus treebank XML (2.1)";
+    public static final String ID = "dkpro-core-uima-inline-xml";
+    public static final String NAME = "Inline XML";
 
     @Override
     public String getId()
@@ -47,21 +49,16 @@ public class PerseusFormatSupport
     }
 
     @Override
-    public boolean isReadable()
+    public boolean isWritable()
     {
         return true;
     }
 
     @Override
-    public boolean isWritable()
-    {
-        return false;
-    }
-
-    @Override
-    public CollectionReaderDescription getReaderDescription(TypeSystemDescription aTSD)
+    public AnalysisEngineDescription getWriterDescription(Project aProject,
+            TypeSystemDescription aTSD, CAS aCAS)
         throws ResourceInitializationException
     {
-        return createReaderDescription(PerseusReader.class, aTSD);
+        return createEngineDescription(InlineXmlWriter.class, aTSD);
     }
 }
